Version 806. AddRef/Release for suspended entities. Removed bson unnecessary destroy. Warnings fixes.
git-svn-id: https://cvet.by:8443/svn/FOnlineSource/Main@1688 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: 0690cfa2fb907ef97def448c5ad0b260563b8640

+++ Source/Crypt.cpp
@@ -378,7 +378,7 @@ uchar* CryptManager::GetCache( const string& data_name, uint& data_len )
     if( r == UNQLITE_NOTFOUND )
         return nullptr;
 
-    uchar* data = new uchar[ size ];
+    uchar* data = new uchar[ (uint) size ];
     r = unqlite_kv_fetch( CacheDb, data_name.c_str(), (int) data_name.length(), data, &size );
     RUNTIME_ASSERT( r == UNQLITE_OK );
